requirements.txt
flask
sqlalchemy
openai
python-dotenv
requests
tweepy
facebook-sdk
linkedin-api
instagram-graph-api
